1. class 시작 후 예외처리 추가 2. Person, Coffee, CoffeeMenu별 swift파일 추가 3. 잔액 부족 기능 수정
